--- src/com/eatthepath/jeospatial/util/package-info.java
@@ -1,7 +1,0 @@
-/**
- * Provides simple implementations of the
- * {@link com.eatthepath.jeospatial.GeospatialPoint} interface and
- * {@code Comparator}s and utility classes for conducting nearest-neighbor
- * searches in {@link com.eatthepath.jeospatial.GeospatialIndex}s.
- */
-package com.eatthepath.jeospatial.util;
